+++ src/request/comments.rs
@@ -88,27 +88,27 @@ impl<'a> GetComments<'a> {
         let mut query = Query::new();
 
         if let Some(sort) = self.sort {
-            query.push("sort", sort.to_string());
+            query.push("sort", &sort.to_string());
         }
 
         if let Some(parent) = self.parent_id {
-            query.push("parent_id", parent.to_string());
+            query.push("parent_id", &parent.to_string());
         }
 
         if let Some(commentable) = self.commentable_id {
-            query.push("commentable_id", commentable.to_string());
+            query.push("commentable_id", &commentable.to_string());
         }
 
         if let Some(commentable) = self.commentable_type.take() {
-            query.push("commentable_type", commentable);
+            query.push("commentable_type", &commentable);
         }
 
         if let Some(cursor) = self.cursor.take() {
-            query.push("cursor[id]", cursor.id.to_string());
-            query.push("cursor[created_at]", cursor.created_at.to_string()); // TODO: Test
+            query.push("cursor[id]", &cursor.id.to_string());
+            query.push("cursor[created_at]", &cursor.created_at.to_string()); // TODO: Test
         }
 
-        let req = Request::from((query, Route::GetComments));
+        let req = Request::new(Route::GetComments).query(query);
 
         Box::pin(self.osu.inner.request(req))
     }