+++ src/request/matches.rs
@@ -64,23 +64,22 @@ impl<'a> GetMatch<'a> {
         let mut query = Query::new();
 
         if let Some(after) = self.after {
-            query.push("after", after.to_string());
+            query.push("after", &after.to_string());
         }
 
         if let Some(before) = self.before {
-            query.push("before", before.to_string());
+            query.push("before", &before.to_string());
         }
 
         if let Some(limit) = self.limit {
-            query.push("limit", limit.to_string());
+            query.push("limit", &limit.to_string());
         }
 
-        let req = Request::from((
-            query,
-            Route::GetMatch {
-                match_id: Some(self.match_id),
-            },
-        ));
+        let route = Route::GetMatch {
+            match_id: Some(self.match_id),
+        };
+
+        let req = Request::new(route).query(query);
 
         Box::pin(self.osu.inner.request(req))
     }
@@ -121,10 +120,10 @@ impl<'a> GetMatches<'a> {
         let mut query = Query::new();
 
         if let Some(cursor) = self.cursor.take() {
-            query.push("cursor[match_id]", cursor.match_id.to_string());
+            query.push("cursor[match_id]", &cursor.match_id.to_string());
         }
 
-        let req = Request::from((query, Route::GetMatch { match_id: None }));
+        let req = Request::new(Route::GetMatch { match_id: None }).query(query);
 
         Box::pin(self.osu.inner.request(req))
     }