+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [Breaking] Removed deprecated `cover_url` field from `User`; use `cover.url` instead
 - [Breaking] `description` field of `Group` is now `Option<String>` instead of `String`
 - [Breaking] Added new `BeatmapsetEvent` variant `OwnerChange` and declared `BeatmapsetEvent` as non-exhaustive
+- [Breaking] `OsuBuilder` no longer accepts a reqwest client since its now using a hyper client
 
 # v0.1.0